# improve performance of find_files_from_load_path