@duckinator 2016-03-11
I'm going through old PRs and trying to figure out what needs done.

1. Is this PR still necessary? (Some PRs have been rendered unnecessary by other changes.)
2. Assuming it is necessary, what is preventing this from being merged, and how should work on it proceed?

It looks like the answer to 2 is a rebase of @tenderlove's work followed by adding a lock around the `Dir.chdir` call. Is that correct?