@segiddins 2016-03-11
A lock around the `Dir.chdir` call is insufficient -- another thread could be changing the directory and obviously wouldn't be using the rubygems mutex.