@djberg96 2017-01-23
Also, I don't know if JRuby ever resolved it:

http://markmail.org/message/jvrzql7g5wj5a5az#query:+page:1+mid:jvrzql7g5wj5a5az+state:results