@duckinator 2018-06-21
This PR has been open for nearly 4 years(!) without finding a way to resolve it, and the current code is not suitable for release due to the use of `Dir.chdir`. 

Given these facts, I'm going to close it.

If you are still dealing with performance problems from this, please create an issue that links to this PR.